Take in the commit bits for the following 8 people:
   brian
   gber
   gleb
   ivoras
   rdivacky
   trhodes
   vanhu
   zont

All of these have not been active for at least 18 months and agree
that they aren't likely to resume FreeBSD work soon, so have aquiesced
to giving up their commit bits.

We haven't had the Grim Reaper working for base commit bits for some
time, so we've a backlog of bits to take in.  More to come.

Approved by:	core (implicit)