rtwn: Fix RTL8188EU cannot associate in STA mode
As with RTL8192EU a user reported that the RTL8188EU has identical
problems. Apply a similar fix to this chipset too.
others it will work fine. On the systems RTL8192EU fails to associate
in STA mode, it works perfectly fine in AP mode. This points to a USB
timing issue when selecting a channel while in STA mode.

PR:		247528
Reported by:	rkoberman@gmail.com
MFC after:	2 weeks
Differential Revision:	https://reviews.freebsd.org/D38854